# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BlueCore SMM Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blog:
    get:
      operationId: getBlogPosts
      tags: [blog]
      summary: Get published blog posts
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostsResponse"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin/manager)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "200":
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /blog/saved/list:
    get:
      operationId: getSavedPosts
      tags: [blog]
      summary: Get saved posts for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Saved posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blog/{id}/save:
    post:
      operationId: toggleSavePost
      tags: [blog]
      summary: Toggle save/unsave a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Save status toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved:
                    type: boolean
                required: [saved]
  /cases:
    get:
      operationId: getCaseStudies
      tags: [cases]
      summary: Get case studies
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Case studies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStudy"
    post:
      operationId: createCaseStudy
      tags: [cases]
      summary: Create a case study (admin/manager)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyInput"
      responses:
        "201":
          description: Case study created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
  /cases/{slug}:
    get:
      operationId: getCaseStudy
      tags: [cases]
      summary: Get a single case study
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
  /cases/{id}:
    put:
      operationId: updateCaseStudy
      tags: [cases]
      summary: Update a case study
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyInput"
      responses:
        "200":
          description: Case study updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
    delete:
      operationId: deleteCaseStudy
      tags: [cases]
      summary: Delete a case study (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /banners:
    get:
      operationId: getBanners
      tags: [banners]
      summary: Get active banners
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: lang
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Active banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Banner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/admin/all:
    get:
      operationId: getAllBanners
      tags: [banners]
      summary: Get all banners (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
  /banners/{id}:
    put:
      operationId: updateBanner
      tags: [banners]
      summary: Update a banner (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "200":
          description: Banner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /banners/{id}/click:
    post:
      operationId: trackBannerClick
      tags: [banners]
      summary: Track a banner click
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Click tracked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MultiLangContent:
      type: object
      properties:
        uz:
          type: object
        ru:
          type: object
        en:
          type: object
    BlogPostInput:
      type: object
      properties:
        slug:
          type: string
        status:
          type: string
        category:
          type: string
        image:
          type: string
        content:
          $ref: "#/components/schemas/MultiLangContent"
      required: [slug, content]
    BlogPost:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        status:
          type: string
        category:
          type: string
        image:
          type: string
        authorId:
          type: number
        content:
          $ref: "#/components/schemas/MultiLangContent"
        views:
          type: number
        publishedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, slug, status, content]
    BlogPostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required: [posts, total, page, limit]
    CaseStudyInput:
      type: object
      properties:
        slug:
          type: string
        client:
          type: string
        platform:
          type: string
        beforeImage:
          type: string
        afterImage:
          type: string
        metrics:
          type: object
        status:
          type: string
        content:
          $ref: "#/components/schemas/MultiLangContent"
        featured:
          type: boolean
        order:
          type: number
      required: [slug, client]
    CaseStudy:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        client:
          type: string
        platform:
          type: string
        beforeImage:
          type: string
        afterImage:
          type: string
        metrics:
          type: object
        status:
          type: string
        content:
          $ref: "#/components/schemas/MultiLangContent"
        featured:
          type: boolean
        order:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, slug, client, status]
...
    BannerInput:
      type: object
      properties:
        type:
          type: string
        active:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
        targetLang:
          type: string
        targetAudience:
          type: string
        abVariant:
          type: string
        content:
          type: object
      required: [content]
    Banner:
      type: object
      properties:
        id:
          type: number
        type:
          type: string
        active:
          type: boolean
        startDate:
          type: string
        endDate:
          type: string
        targetLang:
          type: string
        targetAudience:
          type: string
        abVariant:
          type: string
        content:
          type: object
        clicks:
          type: number
        impressions:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, type, active, content]
...