# lib/api-zod/src/generated/types/blogPostsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueCore SMM Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPost } from "./blogPost";

export interface BlogPostsResponse {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
}