# lib/api-zod/src/generated/types/multiLangContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BlueCore SMM Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { MultiLangContentEn } from "./multiLangContentEn";
import type { MultiLangContentRu } from "./multiLangContentRu";
import type { MultiLangContentUz } from "./multiLangContentUz";

export interface MultiLangContent {
  uz?: MultiLangContentUz;
  ru?: MultiLangContentRu;
  en?: MultiLangContentEn;
}